pyproject.toml
[project]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = ""
authors = [
    {name = "Volodymyr Popov",email = "smartappsvp@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "fastapi[standard] (>=0.115.11,<0.116.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "redis (>=5.2.1,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
